+++ lib/src/geo_static.dart
@@ -2,10 +2,10 @@ import 'package:flutter_osm_plugin/src/geo_point.dart';
 import 'package:flutter_osm_plugin/src/marker.dart';
 
 class StaticPositionGeoPoint{
-
+  final String id;
   final MarkerIcon markerIcon;
   final List<GeoPoint> geoPoints;
 
-  StaticPositionGeoPoint(this.markerIcon,this.geoPoints);
+  StaticPositionGeoPoint(this.id,this.markerIcon,this.geoPoints);
 
 }